+++ onboarding-enabler-python-sample-app/src/pythonSwagger.json
