+++ onboarding-enabler-python-sample-app/src/service-configuration.yml
@@ -40,7 +40,7 @@ instance:
     apiml.service.description: 'Sample API services to demonstrate Python Onboarding Enabler'
 
 ssl:
-  certificate: ../keystore/localhost/localhost.keystore.cer
-  keystore: ../keystore/localhost/localhost.keystore.key
-  caFile: ../keystore/localhost/localhost.pem
+  certificate: ../../keystore/localhost/localhost.keystore.cer
+  keystore: ../../keystore/localhost/localhost.keystore.key
+  caFile: ../../keystore/localhost/localhost.pem
   keyPassword: password
